parsers: removed period parsing from index
