+++ src/parsers/manifest/dash/index.ts
@@ -14,8 +14,8 @@
  * limitations under the License.
  */
 
-import parseMPD from "../../../parsers/manifest/dash/node_parsers";
 import { IParsedManifest } from "../types";
+import parseMPD from "./parseMPD";
 
 /**
  * @param {Document} manifest - Original manifest as returned by the server
@@ -28,7 +28,7 @@ export default function parseFromDocument(
 ) : IParsedManifest {
   const root = document.documentElement;
   if (!root || root.nodeName !== "MPD") {
-    throw new Error("document root should be MPD");
+    throw new Error("DASH Parser: document root should be MPD");
   }
   return parseMPD(root, uri);
 }
