+++ src/parsers/manifest/dash/indexes/template.ts
@@ -91,7 +91,7 @@ export interface ITemplateIndexIndexArgument {
 export interface ITemplateIndexContextArgument {
   periodStart : number; // Start of the period concerned by this
                         // RepresentationIndex, in seconds
-  representationURL : string; // Base URL for the Representation concerned
+  representationBaseURL : string; // Base URL for the Representation concerned
   representationId? : string; // ID of the Representation concerned
   representationBitrate? : number; // Bitrate of the Representation concerned
 }
@@ -110,7 +110,7 @@ export default class TemplateRepresentationIndex implements IRepresentationIndex
   ) {
     const {
       periodStart,
-      representationURL,
+      representationBaseURL,
       representationId,
       representationBitrate,
     } = context;
@@ -128,7 +128,7 @@ export default class TemplateRepresentationIndex implements IRepresentationIndex
       indexTimeOffset,
       initialization: index.initialization && {
         mediaURL: createIndexURL(
-          representationURL,
+          representationBaseURL,
           index.initialization.media,
           representationId,
           representationBitrate
@@ -136,7 +136,7 @@ export default class TemplateRepresentationIndex implements IRepresentationIndex
         range: index.initialization.range,
       },
       mediaURL: createIndexURL(
-        representationURL,
+        representationBaseURL,
         index.media,
         representationId,
         representationBitrate
